fix: move chunker utils tests
License: MIT
